requirements.txt
torch>=2.2
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
plotly>=5.15
tqdm>=4.65
joblib>=1.3
umap-learn>=0.5.5